WHISPER_AI/requirements.txt
streamlit
openai-whisper
librosa
wave
mimetypes
tqdm